requirements.txt
qrcode
pillow